Add a real test
